scripts/gk-detector/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
unidecode>=1.3.6
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
rich>=13.0.0
pytest>=7.4.0
